Use dynamic page for country page
