+++ countries/argentina.json
@@ -1,0 +1,8 @@
+{
+    "name": "Argentina",
+    "language": "Spanish",
+    "continent": "South America",
+    "population": "44,938,712 million",
+    "currency": "Argentine peso",
+    "capital": "Buenos Aires"
+}
