+++ countries/brazil.json
@@ -1,0 +1,8 @@
+{
+    "name": "Brazil",
+    "language": "Portuguese (Brazilian)",
+    "continent": "South America",
+    "population": "210,147,125 million",
+    "currency": "Brazilian real",
+    "capital": "Brasilia"
+}
