+++ countries/canada.json
@@ -1,0 +1,8 @@
+{
+    "name": "Canada",
+    "language": "English, French",
+    "continent": "North America",
+    "population": "38,048,738 million",
+    "currency": "Canadian dollar",
+    "capital": "Ottawa"
+}
