+++ countries/finland.json
@@ -1,0 +1,8 @@
+{
+    "name": "Finland",
+    "language": "Finnish, Swedish",
+    "continent": "Europe",
+    "population": "5,536,146 million",
+    "currency": "Euro",
+    "capital": "Helsinki"
+}
