+++ countries/indonesia.json
@@ -1,0 +1,8 @@
+{
+    "name": "Indonesia",
+    "language": "Bahasa indonesia",
+    "continent": "Asia",
+    "population": "270,203,917 million",
+    "currency": "Indonesian rupiah",
+    "capital": "Jakarta"
+}
