+++ countries/morocco.json
@@ -1,0 +1,8 @@
+{
+    "name": "Morocco",
+    "language": "Arabic, Berber",
+    "continent": "Africa",
+    "population": "37,112,080 million",
+    "currency": "Moroccan dirham",
+    "capital": "Rabat"
+}
