+++ countries/poland.json
@@ -1,0 +1,8 @@
+{
+    "name": "Poland",
+    "language": "Polish",
+    "continent": "Europe",
+    "population": "38,268,000 million",
+    "currency": "Złoty",
+    "capital": "Warsaw"
+}
