+++ countries/united-kingdom.json
@@ -1,0 +1,8 @@
+{
+    "name": "United Kingdom",
+    "language": "English",
+    "continent": "Europe",
+    "population": "63,182,178 million",
+    "currency": "Poundsterling",
+    "capital": "London"
+}
