--- pages/country/argentina.tsx
@@ -1,14 +1,0 @@
-import CountryInfo from "../../components/CountryInfo";
-
-export default function argentina() {
-    return (
-        <CountryInfo 
-            name="Argentina"
-            language="Spanish"
-            continent="South America"
-            population="44,938,712 million"
-            currency="Argentine peso"
-            capital="Buenos Aires"
-        />
-    )
-}
