--- pages/country/brazil.tsx
@@ -1,14 +1,0 @@
-import CountryInfo from "../../components/CountryInfo";
-
-export default function brazil() {
-    return (
-        <CountryInfo 
-            name="Brazil"
-            language="Portuguese (Brazilian)"
-            continent="South America"
-            population="210,147,125 million"
-            currency="Brazilian real"
-            capital="Brasilia"
-        />
-    )
-}
