--- pages/country/canada.tsx
@@ -1,14 +1,0 @@
-import CountryInfo from "../../components/CountryInfo";
-
-export default function canada() {
-    return (
-        <CountryInfo 
-            name="Canada"
-            language="English, French"
-            continent="North America"
-            population="38,048,738 million"
-            currency="Canadian dollar"
-            capital="Ottawa"
-        />
-    )
-}
