--- pages/country/finland.tsx
@@ -1,14 +1,0 @@
-import CountryInfo from "../../components/CountryInfo";
-
-export default function finland() {
-    return (
-        <CountryInfo 
-            name="Finland"
-            language="Finnish, Swedish"
-            continent="Europe"
-            population="5,536,146 million"
-            currency="Euro"
-            capital="Helsinki"
-        />
-    )
-}
