--- pages/country/indonesia.tsx
@@ -1,14 +1,0 @@
-import CountryInfo from "../../components/CountryInfo";
-
-export default function indonesia() {
-    return (
-        <CountryInfo 
-            name="Indonesia"
-            language="Bahasa indonesia"
-            continent="Asia"
-            population="270,203,917 million"
-            currency="Indonesian rupiah"
-            capital="Jakarta"
-        />
-    )
-}
