--- pages/country/morocco.tsx
@@ -1,14 +1,0 @@
-import CountryInfo from "../../components/CountryInfo";
-
-export default function morocco() {
-    return (
-        <CountryInfo 
-            name="Morocco"
-            language="Arabic, Berber"
-            continent="Africa"
-            population="37,112,080 million"
-            currency="Moroccan dirham"
-            capital="Rabat"
-        />
-    )
-}
