--- pages/country/poland.tsx
@@ -1,14 +1,0 @@
-import CountryInfo from "../../components/CountryInfo";
-
-export default function poland() {
-    return (
-        <CountryInfo 
-            name="Poland"
-            language="Polish"
-            continent="Europe"
-            population="38,268,000 million"
-            currency="Złoty"
-            capital="Warsaw"
-        />
-    )
-}
