--- pages/country/united-kingdom.tsx
@@ -1,14 +1,0 @@
-import CountryInfo from "../../components/CountryInfo";
-
-export default function unitedKingdom() {
-    return (
-        <CountryInfo 
-            name="United Kingdom"
-            language="English"
-            continent="Europe"
-            population="63,182,178 million"
-            currency="Poundsterling"
-            capital="London"
-        />
-    )
-}
